Fixed bug that caused the code to not work on systems where "unsigned long" is not-32 bit. Thanks Mark Hasemeyer.

+++ src/TinyTinyJPG.h
@@ -46,7 +46,7 @@ namespace ttjpg {
   /* Unsigned types */
   typedef unsigned char uint8;
   typedef unsigned short uint16;
-  typedef unsigned long uint32;
+  typedef unsigned int uint32;
   
   /* Error codes */
   enum ErrorCode {
